Reorganize files to allow multiple client implementations
