+++ client/koclient.html
@@ -40,6 +40,6 @@
   <script src="bower_components/jeditable/jquery.jeditable.js"></script>
   <script src="https://cdn.rawgit.com/chbrown/rfc6902/v1.0.6/rfc6902.js"></script>
   <script src="bower_components/knockout/dist/knockout.js"></script>
-  <script src="lib/scripts.js"></script>
+  <script src="lib/koclient.js"></script>
 </body>
 </html>
