--- src/_tsd.d.ts
@@ -1,3 +1,0 @@
-/// <reference path="../typings/jquery/jquery.d.ts" />
-/// <reference path="../typings/knockout/knockout.d.ts" />
-/// <reference path="../typings/socket.io-client/socket.io-client.d.ts" />
