+++ src/custom-typings/defaults.d.ts
@@ -1,0 +1,3 @@
+// Generic dictionary syntactic sugar
+declare type Dictionary<TValue> = { [id: string]: TValue; }
+
