+++ src/custom-typings/jquery-jeditable.d.ts
@@ -1,0 +1,5 @@
+// bower_components/jeditable/jquery.jeditable.js
+interface JQuery {
+  editable(fn: (value: any, params: any) => any, options: any): JQuery;
+  editable(dest: string): JQuery;
+}
