--- src/main.ts
@@ -1,8 +1,0 @@
-import DragNDrop = require("./kobindings/dragndrop");
-DragNDrop.register();
-
-import JEditable = require("./kobindings/jeditable");
-JEditable.register();
-
-import App = require("./app");
-App.start();
