+++ tsd.json
@@ -3,7 +3,6 @@
   "repo": "borisyankov/DefinitelyTyped",
   "ref": "master",
   "path": "typings",
-  "bundle": "src/_tsd.d.ts",
   "installed": {
     "knockout/knockout.d.ts": {
       "commit": "419793fb796b0e5c1ed11e018ad28475de0c3735"
